1.修复服务器重复打开关闭时 发送和接收队列没有被重置问题, 2.修复文件数据的int值不够用问题, 改成long类型 3.更新当客户端退出,删除.stream临时文件

+++ GameDesigner/Network/core/Share/FileData.cs
@@ -22,7 +22,7 @@ namespace Net.Share
         /// <summary>
         /// 文件总长度
         /// </summary>
-        public int Length;
+        public long Length;
         /// <summary>
         /// 实际文件写入对象
         /// </summary>
